(PUP-4363) Move splay handling to utility module
This is to allow it to be shared between the agent and other
applications.
