+++ spec/unit/agent_spec.rb
@@ -234,44 +234,6 @@ describe Puppet::Agent do
     end
   end
 
-  describe "when splaying" do
-    before do
-      Puppet[:splay] = true
-      Puppet[:splaylimit] = "10"
-    end
-
-    it "should do nothing if splay is disabled" do
-      Puppet[:splay] = false
-      @agent.expects(:sleep).never
-      @agent.splay
-    end
-
-    it "should do nothing if it has already splayed" do
-      @agent.expects(:splayed?).returns true
-      @agent.expects(:sleep).never
-      @agent.splay
-    end
-
-    it "should log that it is splaying" do
-      @agent.stubs :sleep
-      Puppet.expects :info
-      @agent.splay
-    end
-
-    it "should sleep for a random portion of the splaylimit plus 1" do
-      Puppet[:splaylimit] = "50"
-      @agent.expects(:rand).with(51).returns 10
-      @agent.expects(:sleep).with(10)
-      @agent.splay
-    end
-
-    it "should mark that it has splayed" do
-      @agent.stubs(:sleep)
-      @agent.splay
-      expect(@agent).to be_splayed
-    end
-  end
-
   describe "when checking execution state" do
     describe 'with regular run status' do
       before :each do
